requirements.txt
click
pillow
pillow-heif
piexif